Move store_bids_tree
